Move game logic and AI to separate packages
Do some further code cleanups
