+++ app/src/main/java/com/cyberbot/checkers/game/logic/GameEndReason.java
@@ -1,4 +1,4 @@
-package com.cyberbot.checkers.game;
+package com.cyberbot.checkers.game.logic;
 
 public enum GameEndReason {
     // Draws
