+++ app/src/main/java/com/cyberbot/checkers/game/logic/PieceType.java
@@ -1,4 +1,4 @@
-package com.cyberbot.checkers.game;
+package com.cyberbot.checkers.game.logic;
 
 public enum PieceType {
     ORDINARY,
