+++ app/src/main/java/com/cyberbot/checkers/ui/animator/MoveScaleAnimator.kt
@@ -2,7 +2,7 @@ package com.cyberbot.checkers.ui.animator
 
 import android.animation.AnimatorSet
 import android.animation.ValueAnimator
-import com.cyberbot.checkers.game.GridEntry
+import com.cyberbot.checkers.game.logic.GridEntry
 
 class MoveScaleAnimator(singleCellSize: Float, sequential: Boolean = false) :
     PieceAnimator(singleCellSize, sequential) {
