+++ app/src/main/java/com/cyberbot/checkers/ui/animator/ScaleAnimator.kt
@@ -1,7 +1,7 @@
 package com.cyberbot.checkers.ui.animator
 
 import android.animation.ValueAnimator
-import com.cyberbot.checkers.game.GridEntry
+import com.cyberbot.checkers.game.logic.GridEntry
 
 class ScaleAnimator(singleCellSize: Float, sequential: Boolean = false) :
     PieceAnimator(singleCellSize, sequential) {
